* Implemented DAG for transferring packages data from SFTP-server (1C) to postgresql.

+++ dags/upload_packages_sftp_1c_to_pg/dag.py
@@ -3,24 +3,17 @@ import os
 
 from airflow import DAG
 from airflow.providers.sftp.hooks.sftp import SFTPHook
-from airflow.sdk import task
+from airflow.sdk import task, teardown
 from datetime import datetime, UTC
 
-from airflow.utils.trigger_rule import TriggerRule
+from constants import AIRFLOW_TMP_DIRPATH, TZ_MSK
 
-from dags.constants import AIRFLOW_TMP_DIRPATH, TZ_MSK
-
-DEFAULT_ARGS = {
-    "retries": 1,
-    "retry_delay": 30,
-}
 
 with DAG(
     dag_id="upload_packages_sftp_1c_to_pg",
     start_date=datetime(2025, 5, 1, tzinfo=TZ_MSK),
     schedule='30 1 * * *',
     catchup=False,
-    default_args=DEFAULT_ARGS,
 ) as dag:
     @task
     def download_task():
@@ -36,24 +29,25 @@ with DAG(
 
     @task
     def transform_task(source_fp: str):
-        from dags.upload_packages_sftp_1c_to_pg.libs.transform import transform
+        from upload_packages_sftp_1c_to_pg.libs.transform import transform
+
         transformed_fp = f"{source_fp}_t.csv"
         transform(source_fp, transformed_fp)
         return transformed_fp
 
     @task
     def upload_task(transformed_fp: str):
-        from dags.upload_packages_sftp_1c_to_pg.libs.upload import PgPackagesHook
+        from upload_packages_sftp_1c_to_pg.libs.upload import PgPackagesHook
 
         pg = PgPackagesHook('pg_prod', transformed_fp)
         pg.create_table()
         pg.clear_table()
         pg.import_data()
 
-    @task(trigger_rule=TriggerRule.ALL_DONE)
+    @teardown
     def cleanup_task(fpaths: list[str]):
         for fp in fpaths:
-            if os.path.exists(fp):
+            if fp and os.path.exists(fp):
                 os.remove(fp)
                 logging.info("File %s removed.", fp)
